2fa test rolled out

+++ html_api/settings/twofactor/disable.php
@@ -10,4 +10,10 @@ header("access-control-allow-credentials: true");
 header('Content-Type: application/json');
 
 $userid = $user->id;
+
+//feature tester locked
+if (!inFeatureTesterGroup($userid)) {
+	die(http_response_code(401));
+}
+
 echo json_encode(array("success" => deleteUser2FA($userid)));
